# applyOpacity simd for image too, 10x faster as well 