2560x1440 image
before:
`applyOpacity ....................... 7.058 ms      7.172 ms    ±0.116   x695`
after:
`applyOpacity ....................... 0.759 ms      0.779 ms    ±0.038  x1000`